fix verification and validation
